WEBP: remove cast, since older webp expects a uint32 and not a size_t, which raises warning about implicit downcasting. We check above that the result of the multiplication is not larger than INT_MAX
git-svn-id: https://svn.osgeo.org/gdal/trunk@38836 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/webp/webpdataset.cpp
@@ -314,14 +314,14 @@ CPLErr WEBPDataset::Uncompress()
             pabyCompressed,
             static_cast<uint32_t>( nSize ),
             static_cast<uint8_t*>( pabyUncompressed),
-            static_cast<size_t>(nRasterXSize) * nRasterYSize * nBands,
+            nRasterXSize * nRasterYSize * nBands,
             nRasterXSize * nBands );
     else
         pRet = WebPDecodeRGBInto(
             pabyCompressed,
             static_cast<uint32_t>( nSize ),
             static_cast<uint8_t*>( pabyUncompressed ),
-            static_cast<size_t>(nRasterXSize) * nRasterYSize * nBands,
+            nRasterXSize * nRasterYSize * nBands,
             nRasterXSize * nBands );
 
     VSIFree(pabyCompressed);
